bugfix: add/rename `TempoMap.Sync(…)` to `Finalize(…)`
this is logically awkward...

and add support for a 'soft' finalize for cases where
the first Tempo received is at pulse > 0 — as we need
to insert a 'default' tempo at pulse 0 with max = n;
where of course n > 0 and is our first received tempo.

this is specifically due to the fact that we're
attempting to calculate a number of seconds elapsed
at each parsed midi message event.  Additionally, the
call to `TempoMap.Finalize()` can only (in this case)
be done after we know the length of each track after
parsing all tracks.

So if you manage to run into a case where you attempt
using the TempoMap before all tracks have been read
(e.g. before finalization of the tempo-map occurs)
BUT you have the TempoMap (Track 0) [<--- you see
there is a logical error somewhere...] such as
seen in smfio.view where our first tempo in the map
is received at pulse 1 (i.e. > 0) --- our TempoMap
fails to find a working tempo; hence, we need
a 'soft' Finalize.

so me thinks some simplifica/consolida-tion and/or
clarity around this may be found in the future.